refactor: ticketId로 Result 화면 로드 기능 구현

+++ app/src/main/java/woowacourse/movie/model/ticket/ReservationTicket.kt
@@ -9,7 +9,7 @@ import java.io.Serializable
 @Entity(tableName = "reservationTicket")
 data class ReservationTicket(
     @PrimaryKey(autoGenerate = true)
-    val ticketId: Int = 0,
+    val ticketId: Long = 0,
     val movieId: Int,
     val theaterId: Int,
     val movieTitle: String,
